# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chainsmith API",
    "description": "Headless Rust API for the Chainsmith Flesh and Blood deck builder.",
    "license": {
      "name": "PolyForm-Noncommercial-1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/validate": {
      "post": {
        "operationId": "validateDeck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation completed; check `valid` and `violations`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or unsupported format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CardId": {
        "type": "string"
      },
...
      "Deck": {
        "type": "object",
        "required": [
          "format",
          "hero",
          "pool",
          "loadouts"
        ],
        "properties": {
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "hero": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "loadouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loadout"
            }
          },
          "pool": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolEntry"
            }
          }
        },
        "additionalProperties": false
      },
...
      "EquipmentLoadout": {
        "type": "object",
        "properties": {
          "arms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "chest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "head": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "legs": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "mainHand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId",
                "description": "Main-hand slot. Holds a weapon (1H or 2H). A 2H weapon here occupies\nboth hand slots; see `off_hand` for the Companion exception."
              }
            ]
          },
          "offHand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId",
                "description": "Off-hand slot. Holds either a 1H weapon, an off-hand equipment piece\n(shield, Trade Routes, etc.), or — when `main_hand` carries a 2H\nweapon — only a Companion-subtyped off-hand equipment piece."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Stable, machine-readable error codes. Clients are allowed to switch on\nthese. Adding a variant is additive; renaming or removing one is a\nbreaking change in production phase (see `.claude/rules/api-contract.md`).",
        "enum": [
          "database_error",
          "unsupported_format",
          "not_found",
          "unauthorized",
          "forbidden",
          "auth_not_configured"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "FormatId": {
        "type": "string",
        "enum": [
          "classic_constructed",
          "blitz",
          "commoner",
          "silver_age",
          "living_legend",
          "ultimate_pit_fight"
        ]
      },
...
      "Loadout": {
        "type": "object",
        "required": [
          "name",
          "deckCards",
          "equipment"
        ],
        "properties": {
          "deckCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadoutEntry"
            }
          },
          "equipment": {
            "$ref": "#/components/schemas/EquipmentLoadout"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoadoutEntry": {
        "type": "object",
        "required": [
          "printing",
          "quantity"
        ],
        "properties": {
          "printing": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "PoolEntry": {
        "type": "object",
        "required": [
          "printing",
          "quantity"
        ],
        "properties": {
          "printing": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "PrintingId": {
        "type": "string"
      },
...
      "ValidateRequest": {
        "type": "object",
        "required": [
          "deck"
        ],
        "properties": {
          "date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Validation date. Used to evaluate time-evolving rules (banned list,\nLiving Legend retirement). Defaults to today UTC if omitted."
          },
          "deck": {
            "$ref": "#/components/schemas/Deck"
          }
        },
        "additionalProperties": false
      },
      "ValidationResponse": {
        "type": "object",
        "required": [
          "valid",
          "format",
          "date",
          "violations"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "valid": {
            "type": "boolean",
            "description": "True iff `violations` is empty."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        }
      },
      "Violation": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ViolationCode"
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ViolationDetails"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ViolationCode": {
        "type": "string",
        "description": "Stable codes clients are allowed to switch on.\n\nAdding a variant is additive. Renaming or removing one is a breaking\nchange in the production phase (see `.claude/rules/api-contract.md`).",
        "enum": [
          "hero_printing_not_found_in_catalog",
          "hero_not_eligible_for_format",
          "hero_living_legend_retired",
          "hero_missing_hero_type",
          "pool_size_above_max",
          "deck_size_below_min",
          "deck_size_above_max",
          "copy_limit_exceeded",
          "restricted_copy_limit_exceeded",
          "card_not_found_in_catalog",
          "printing_not_found_in_catalog",
          "card_not_eligible_for_format",
          "card_banned",
          "supertype_mismatch",
          "specialization_mismatch",
          "loadout_printing_not_in_pool",
          "loadout_quantity_exceeds_pool",
          "equipment_slot_wrong",
          "equipment_not_in_pool",
          "weapon_config_invalid"
        ]
      },
      "ViolationDetails": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "card_id",
              "name",
              "kind"
            ],
            "properties": {
              "card_id": {
                "$ref": "#/components/schemas/CardId"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "card"
                ]
              },
              "name": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "printing_id",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "printing"
                ]
              },
              "printing_id": {
                "$ref": "#/components/schemas/PrintingId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "card_id",
              "found",
              "allowed",
              "kind"
            ],
            "properties": {
              "allowed": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "card_id": {
                "$ref": "#/components/schemas/CardId"
              },
              "found": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "kind": {
                "type": "string",
                "enum": [
                  "quantity"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "found",
              "min",
              "kind"
            ],
            "properties": {
              "found": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "kind": {
                "type": "string",
                "enum": [
                  "deck_size"
                ]
              },
              "max": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              },
              "min": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "required": [
              "found",
              "max",
              "kind"
            ],
            "properties": {
              "found": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "kind": {
                "type": "string",
                "enum": [
                  "pool_size"
                ]
              },
              "max": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "required": [
              "printing_id",
              "expected_slot",
              "kind"
            ],
            "properties": {
              "actual_slot": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "expected_slot": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "equipment_slot"
                ]
              },
              "printing_id": {
                "$ref": "#/components/schemas/PrintingId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "loadout",
              "printing_id",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "loadout_card"
                ]
              },
              "loadout": {
                "type": "string"
              },
              "printing_id": {
                "$ref": "#/components/schemas/PrintingId"
              }
            }
          }
        ]
      },
...
    }
  }
}